requirements.txt
together>=0.2.5
tiktoken>=0.5.0
python-dotenv>=1.0.0
typing-extensions>=4.5.0
flask>=2.0.0
PyMuPDF  # for PDF parsing 
tqdm
transformers >= 4.17.0.dev0
accelerate
datasets >= 1.8.0
sentencepiece != 0.1.92
protobuf
rouge-score
nltk
py7zr
evaluate
prettytable
openai
requests
huggingface_hub

numpy<2.0
matplotlib
pydantic
gym
bs4
beautifulsoup4
python-leetcode
spacy
openpyxl
